requirements.txt
Pillow~=10.1.0
pyautogui >= 0.9

PyScreeze~=0.1.30